+++ app/src/main/java/com/studhub/app/presentation/profile/components/EditProfileForm.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.studhub.app.R
+import com.studhub.app.core.utils.isValidSwissPhoneNumber
 import com.studhub.app.presentation.ui.common.button.BasicFilledButton
 import com.studhub.app.presentation.ui.common.input.BasicTextField
 import com.studhub.app.presentation.ui.common.input.ImagePicker
@@ -76,9 +77,4 @@ fun EditProfileForm(
     }
 }
 
-fun isValidSwissPhoneNumber(phoneNumber: String): Boolean {
-    // Swiss phone numbers can start with +41, 0041, 0 or no prefix followed by 9 digits
-    val regex = Regex("^((\\+41|0041|0)?[1-9]\\d{1}\\s?\\d{3}\\s?\\d{2}\\s?\\d{2})$")
-    return regex.matches(phoneNumber)
-}
