+++ app/src/main/res/values/strings_profile.xml
@@ -12,4 +12,5 @@
     <string name="profile_edit_form_label_username">User Name</string>
     <string name="profile_edit_form_label_phone">Phone Number</string>
     <string name="profile_edit_form_btn_save">Save information</string>
+    <string name="profile_edit_form_error_invalid_phone">Please enter a valid Swiss phone number</string>
 </resources>
